Rename `contract_address_identifier` -> `felt_or_id`
