+++ crates/sncast/src/starknet_commands/utils/mod.rs
@@ -20,7 +20,7 @@ use crate::{
 };
 
 pub mod class_hash;
-pub mod contract_address_identifier;
+pub mod felt_or_id;
 pub mod serialize;
 
 #[derive(Args)]
